* ranges updated to be non-inclusive-endian * Frame_t has to be handled differently now: using values that are invalid now causes an exception * frame_rate changed to uint32_t because that makes sense anyway * PairingGraph only handles fish ids now, not pointers

+++ Application/src/grabber/pvinfo.cpp
@@ -499,11 +499,11 @@ int main(int argc, char**argv) {
             track::Tracker::auto_calculate_parameters(video, be_quiet);
         }
         
-        if(SETTING(frame_rate).value<int>() == 0) {
+        if(SETTING(frame_rate).value<uint32_t>() == 0) {
             if(!SETTING(quiet))
                 FormatWarning("frame_rate == 0, calculating from frame tdeltas.");
             video.generate_average_tdelta();
-            SETTING(frame_rate) = max(1, int(video.framerate()));
+            SETTING(frame_rate) = (uint32_t)max(1, int(video.framerate()));
         }
         
         Output::Library::Init();
